refactor: organize experiments sub-module for convenient imports (#3687)
